% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert mnda graph data to igraph}
\usage{
as_igraph(mnda.graph, edge.threshold = 0)
}
\arguments{
\item{mnda.graph}{mnda graph data}

\item{edge.threshold}{numeric}
}
\value{
igraph object
}
\description{
Convert mnda graph data to igraph
}
\examples{
data = example_data()
graph = as_igraph(mnda.graph = data[["mnda_graph_example"]])

}
