% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{distance}
\alias{distance}
\title{Function to calculate distance between two vectors}
\usage{
distance(x, y, method = "cosine")
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{method}{distance calculation method: cosine (default), dot.prod, euclidian, manhattan, chebyshev, coassociation}
}
\value{
the distance value
}
\description{
Function to calculate distance between two vectors
}
\examples{
x = c(1,2,3)
y = c(6,4,6)
distance(x,y)

}
