% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_data}
\alias{example_data}
\title{Example Data}
\usage{
example_data()
}
\value{
example data as a list: "adj_mat_example", "igraph_example", "mnda_graph_example"
}
\description{
Example Data
}
\examples{
data = example_data()
}
