% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLayerNet_functions.R
\name{mnda_distance_test_isn}
\alias{mnda_distance_test_isn}
\title{Test the embedding distances of local neighbors change between the two conditions for ISNs.}
\usage{
mnda_distance_test_isn(
  Distance,
  y,
  stat.test = "wilcox.test",
  p.adjust.method = "none"
)
}
\arguments{
\item{stat.test}{statistical test used to detect the nodes \code{c("t.test","wilcox.test")} (default: wilcox.test)}

\item{p.adjust.method}{method for adjusting p-value (including methods on \code{p.adjust.methods}).
If set to "none" (default), no adjustment will be performed.}

\item{distance}{a distance list obtained by the \code{mnda_node_distance()} function.}

\item{alpha}{numeric value of significance level (default: 0.05)}
}
\value{
the highly variable nodes
}
\description{
Test the embedding distances of local neighbors change between the two conditions for ISNs.
}
\details{
The adjusted pvalues for each node.
}
\examples{
myNet = network_gen(N_nodes = 50, N_var_nodes = 5, N_var_nei = 40, noise_sd = .01)
graph_data = myNet[["data_graph"]]
embeddingSpaceList = mnda_embedding(graph.data=graph_data, outcome=c(1,2), indv.index = c(1,1), train.rep=5, walk.rep=5)
Dist = mnda_node_distance(embeddingSpaceList)
Result = mnda_node_detection_isn(Dist, y = c(1,2))

}
