% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2LayerNet_functions.R
\name{mnda_node_detection_2layer}
\alias{mnda_node_detection_2layer}
\title{Detecting the nodes whose local neighbors change bweteen the two conditions.}
\usage{
mnda_node_detection_2layer(embeddingSpaceList)
}
\arguments{
\item{embeddingSpaceList}{a list obtained by the \code{mnda_embedding_2layer()} function.}
}
\value{
the highly varibale nodes
}
\description{
Detecting the nodes whose local neighbors change bweteen the two conditions.
}
\examples{
embeddingSpaceList = mnda_embedding_2layer(graph.data)
Nodes = mnda_node_detection_2layer(embeddingSpaceList)

}
