% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_analysis.R
\name{p_val_count}
\alias{p_val_count}
\title{Calculate p.value for x given set of null values using counts}
\usage{
p_val_count(x, null.values, alternative = "two.sided")
}
\arguments{
\item{x}{numeric value}

\item{null.values}{a numeric vector of null distribution samples}

\item{alternative}{alterative test including: \code{"two.sided"} [default],
\code{"greater"}, and  \code{"less"}}
}
\value{
p.value
}
\description{
Calculate p.value for x given set of null values using counts
}
\examples{
p.val = p_val_count(1, 1:100)

}
