% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_analysis.R
\name{p_val_norm}
\alias{p_val_norm}
\title{Calculate p.value for x given a Gaussian null pdf}
\usage{
p_val_norm(x, null.pdf, alternative = "two.sided")
}
\arguments{
\item{x}{numeric value}

\item{null.pdf}{a numeric vector of null distribution samples}

\item{alternative}{alterative test including: \code{"two.sided"} [default],
\code{"greater"}, and  \code{"less"}}
}
\value{
p.value
}
\description{
Calculate p.value for x given a Gaussian null pdf
}
\examples{
p.val = p_val_norm(1, rnorm(1000,0,1))

}
