% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph_plot.R
\name{subgraph_plot}
\alias{subgraph_plot}
\title{Visualization of a subgroup using a circular graph}
\usage{
subgraph_plot(
  graph,
  node_set,
  labels = NULL,
  node_size = 5,
  font_size = 4,
  max_edge_width = 4,
  margin = 2.5
)
}
\arguments{
\item{graph}{an igraph object}

\item{node_set}{the names or indices of the nodes around which the subgroup is plotted.}

\item{labels}{the labels of the nodes to be indicated. Labels should be a named vector if the \code{node_set} consists of the node names.}

\item{node_size}{size of the nodes in plot (default: 5)}

\item{font_size}{font size of labels if available (default: 4)}

\item{max_edge_width}{maximum thickness of the edges in plot (default: 4)}

\item{margin}{the figure margin (default: 2.5)}
}
\value{
nothing to return
}
\description{
Visualization of a subgroup using a circular graph
}
\details{
This function plots a sub-graph given by a set of nodes as circular plot.
the main inputs to the function are: a graph (as an igraph object) and a set of nodes
(e.g. highly variable nodes) around which the subgroup is calculated.
}
\examples{
Labels = names(V(graph))
subgraph_plot(graph_plot, Node_set, Labels)

}
