% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{WeightdRandomWalk}
\alias{WeightdRandomWalk}
\title{Weighted Random Walk algorithm}
\usage{
WeightdRandomWalk(graph, startNode, maxStep = 5, node_names = FALSE)
}
\arguments{
\item{graph}{an igraph object}

\item{startNode}{the starting node (i.e. a node name or a node index)}

\item{maxStep}{maximum number steps (default:5)}

\item{node_names}{a list of names for nodes}
}
\value{
The set of nodes passed by the random walker.
}
\description{
Weighted Random Walk algorithm
}
\examples{
data = example_data()
nodePath = WeightdRandomWalk(graph = data[["igraph_example"]], startNode = 1)

}
