% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{as.mnda.graph}
\alias{as.mnda.graph}
\title{Convert adjacency matrix to mnda graph data}
\usage{
as.mnda.graph(adj.list, outcome = NULL)
}
\arguments{
\item{adj.list}{list of adjacency matrices with matching nodes}

\item{outcome}{graph outcomes or graph labels. If NULL, \code{outcome = 1:N_graphs}.}
}
\value{
mnda.graph data
}
\description{
Convert adjacency matrix to mnda graph data
}
\examples{
adj.list = list(Adj_mat_example, Adj_mat_example)
graph.data = as.mnda.graph(adj.list)

}
